requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.1.0
dinov2==0.0.1.dev0
filelock==3.18.0
fonttools==4.56.0
foxglove-schemas-protobuf==0.3.0
frozenlist==1.6.0
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.14.3
lz4==4.4.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mcap==1.2.2
mcap-protobuf-support==0.5.3
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
segment-anything==1.0
setuptools==76.0.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wget==3.2
yarl==1.20.0
zstandard==0.23.0
